gateware/top: fix Aurora user clock constraints
For Aurora 8b10b, the user clock is the actual
user clock from the Aurora core. User data interface
is 32-bit.

For Aurora 64b66b, the user clock is 2x the user clock
from the aurora core, because we want the system design
to have the same interface (32-bit data). So, the clock has
to be doubled by a PLL to cope with the fact that aurora
64b66b uses 64-bit interface.